+++ src/app/(auth)/confirm-reset-password/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import BigButton from '@/components/BigButton';
+import { BigBlueButton } from '@/components/Buttons';
 import { H2, H4 } from '@/styles/text';
 import { useRouter } from 'next/navigation';
 import { SpacerDiv } from '../styles';
@@ -11,9 +11,9 @@ export default function ConfirmResetPassword() {
   return (
     <SpacerDiv>
       <H2>Your password has been reset.</H2>
-      <BigButton type="button" onClick={() => push('/login')}>
+      <BigBlueButton type="button" onClick={() => push('/login')}>
         <H4 $color="white">Go to Log In</H4>
-      </BigButton>
+      </BigBlueButton>
     </SpacerDiv>
   );
 }
