+++ src/app/(auth)/email-verified/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { SpacerDiv } from '@/app/(auth)/styles';
-import BigButton from '@/components/BigButton';
+import { BigBlueButton } from '@/components/Buttons';
 import { H2 } from '@/styles/text';
 import { useRouter } from 'next/navigation';
 
@@ -11,9 +11,9 @@ export default function EmailVerified() {
   return (
     <SpacerDiv>
       <H2>Your email has been verified!</H2>
-      <BigButton type="button" onClick={() => push('/onboarding/roles')}>
+      <BigBlueButton type="button" onClick={() => push('/onboarding/roles')}>
         Go to Onboarding
-      </BigButton>
+      </BigBlueButton>
     </SpacerDiv>
   );
 }
