+++ src/app/(auth)/forgot-password/page.tsx
@@ -2,8 +2,7 @@
 
 import supabase from '@/api/supabase/createClient';
 import { HorizontalDiv, SpacerDiv } from '@/app/(auth)/styles';
-import BigButton from '@/components/BigButton';
-import Button from '@/components/Button';
+import { BigBlueButton, Button } from '@/components/Buttons';
 import TextInput from '@/components/TextInput/index';
 import COLORS from '@/styles/colors';
 import { H1, H2, H4, P } from '@/styles/text';
@@ -41,9 +40,9 @@ export default function ForgotPassword() {
             value={email}
             setValue={setEmail}
           />
-          <BigButton type="button" onClick={sendPasswordResetLink}>
+          <BigBlueButton type="button" onClick={sendPasswordResetLink}>
             <H4 $color="white">Send link to email</H4>
-          </BigButton>
+          </BigBlueButton>
         </>
       )}
       {emailSentCount > 0 && (
