+++ src/app/onboarding/basic-information/page.tsx
@@ -1,12 +1,11 @@
 'use client';
 
 import BigDataDropdown from '@/components/BigDataDropdown';
-import { BigButton, BigLinkButton } from '@/components/Button';
+import { BigBlueButton, BigLinkButton } from '@/components/Buttons';
 import { FormControl, FormField, FormItem, FormLabel } from '@/components/Form';
 import TextInput from '@/components/TextInput';
 import { cities } from '@/data/citiesAndStates';
 import { languages } from '@/data/languages';
-import COLORS from '@/styles/colors';
 import { Flex } from '@/styles/containers';
 import { H1Centered } from '@/styles/text';
 import { useGuardedOnboarding } from '@/utils/hooks';
@@ -260,15 +259,9 @@ export default function Page() {
               Back
             </BigLinkButton>
           )}
-          <BigButton
-            type="submit"
-            disabled={isEmpty}
-            $primaryColor={COLORS.blueMid}
-            $secondaryColor={COLORS.blueDark}
-            $tertiaryColor={COLORS.blueDarker}
-          >
+          <BigBlueButton type="submit" disabled={isEmpty}>
             Continue
-          </BigButton>
+          </BigBlueButton>
         </Flex>
       </FormDiv>
     </FormProvider>
