+++ src/app/onboarding/legal-experience/page.tsx
@@ -1,12 +1,11 @@
 'use client';
 
 import BigDataDropdown from '@/components/BigDataDropdown';
-import { BigButton, BigLinkButton } from '@/components/Button';
+import { BigBlueButton, BigLinkButton } from '@/components/Buttons';
 import { FormControl, FormField, FormItem, FormLabel } from '@/components/Form';
 import RadioGroup from '@/components/RadioGroup';
 import TextInput from '@/components/TextInput';
 import { states } from '@/data/citiesAndStates';
-import COLORS from '@/styles/colors';
 import { Flex } from '@/styles/containers';
 import { H1Centered } from '@/styles/text';
 import { formatTruthy } from '@/utils/helpers';
@@ -142,15 +141,9 @@ export default function Page() {
               Back
             </BigLinkButton>
           )}
-          <BigButton
-            type="submit"
-            disabled={isEmpty}
-            $primaryColor={COLORS.blueMid}
-            $secondaryColor={COLORS.blueDark}
-            $tertiaryColor={COLORS.blueDarker}
-          >
+          <BigBlueButton type="submit" disabled={isEmpty}>
             Continue
-          </BigButton>
+          </BigBlueButton>
         </Flex>
       </FormDiv>
     </FormProvider>
