+++ src/app/onboarding/legal-fellow-experience/page.tsx
@@ -7,8 +7,7 @@ import { FormProvider, useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { FormControl, FormField, FormItem, FormLabel } from '@/components/Form';
 import { Flex } from '@/styles/containers';
-import { BigButton, BigLinkButton } from '@/components/Button';
-import COLORS from '@/styles/colors';
+import { BigBlueButton, BigLinkButton } from '@/components/Buttons';
 import { useRouter } from 'next/navigation';
 import { useGuardedOnboarding } from '@/utils/hooks';
 import { formatTruthy, getCurrentDate, parseDateAlt } from '@/utils/helpers';
@@ -131,15 +130,9 @@ export default function Page() {
               Back
             </BigLinkButton>
           )}
-          <BigButton
-            type="submit"
-            disabled={isEmpty}
-            $primaryColor={COLORS.blueMid}
-            $secondaryColor={COLORS.blueDark}
-            $tertiaryColor={COLORS.blueDarker}
-          >
+          <BigBlueButton type="submit" disabled={isEmpty}>
             Continue
-          </BigButton>
+          </BigBlueButton>
         </Flex>
       </FormDiv>
     </FormProvider>
