+++ src/components/InterestForm/index.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import { upsertInterest } from '@/api/supabase/queries/interest';
-import Button from '@/components/Button';
 import DateInput from '@/components/DateInput';
 import RadioGroup from '@/components/RadioGroup';
 import TextAreaInput from '@/components/TextAreaInput';
@@ -12,6 +11,7 @@ import { Interest, Listing } from '@/types/schema';
 import { useAuth } from '@/utils/AuthProvider';
 import { isValidDate } from '@/utils/helpers';
 import { useEffect, useState } from 'react';
+import { Button } from '../Buttons';
 import * as Styles from './styles';
 
 interface Responses {
